move pressable to separate directory
